refactor(controller): isolate modeldeployment metrics state
